+++ Translate5/MaintenanceCli/Command/StatisticsAggregateCommand.php
@@ -196,8 +196,8 @@ class StatisticsAggregateCommand extends Translate5AbstractCommand
         Zend_Registry::get('logger')
             ->cloneMe('core.db.statistics')
             ->info('E1722', 'Statistics::aggregate - {totalCount} records processed in {duration}', [
-                '{totalCount}' => $totalCount,
-                '{duration}' => $duration,
+                'totalCount' => $totalCount,
+                'duration' => $duration,
             ]);
 
         return self::SUCCESS;