+++ application/modules/editor/Controllers/TaskController.php
@@ -288,6 +288,7 @@ class editor_TaskController extends ZfExtended_RestController
             $this->adjustFilter($this->entity->getFilter()),
             (int) $this->getParam('start', 0),
             (int) $this->getParam('limit', 0),
+            false
         );
 
         $kpi = new editor_Models_KPI(SegmentHistoryAggregationRepository::create());
@@ -2471,7 +2472,7 @@ class editor_TaskController extends ZfExtended_RestController
         }
 
         // get all matching tasks before filtering out non-relevant ones
-        $taskDataList = $this->taskViewDataProvider->getTaskList($this->authenticatedUser, $filter);
+        $taskDataList = $this->taskViewDataProvider->getTaskList($this->authenticatedUser, $filter, buildTaskView: false);
 
         if (! empty($taskDataList['rows'])) { // tasks re-filtering may be needed
             $taskGuids = array_column($taskDataList['rows'], 'taskGuid');