+++ application/modules/editor/Models/Import/FileParser/Xlf/Exception.php
@@ -42,9 +42,9 @@ class editor_Models_Import_FileParser_Xlf_Exception extends editor_Models_Import
         'E1190' => 'The XML of the XLF file "{fileName} (id {fileId})" is invalid!',
         'E1191' => 'The XLF file "{fileName} (id {fileId})" does not contain any translation relevant segments.',
         'E1194' => 'The file "{file}" contains "{tag}" tags, which are currently not supported! Stop Import.',
-        'E1195' => 'A trans-unit of file "{file}" contains MRK tags other than type=seg, which are currently not supported! Stop Import.',
+        'E1195' => 'A trans-unit of file "{file}" contains MRK tags of type "{mtypeFound}", which are currently not supported! Stop Import.',
         'E1196' => 'Whitespace in text content of file "{file}" can not be cleaned by preg_replace. Error Message: "{pregMsg}". Stop Import.',
-        'E1232' => 'XLF Parser supports only XLIFF Version 1.1 and 1.2, but the imported xliff tag does not match that criteria: {tag}',
+        'E1232' => 'Only XLF 1.1 and 1.2 are natively supported. You can use OKAPI to import XLIFF Version 2 files.',
         'E1363' => 'Unknown XLF tag found: {tag} - can not import that.',
         'E1689' => 'Your file "{file}" contains texts within cdata blocks. This cannot be imported with the standard translate5 xliff import. However, you can map the file format template "xliff 1.2 with cdata" to the file extension xlf and/or xliff in the file format settings of translate5 (and adapt the template further if necessary) and then import your file again with this template.',
     ];